move clean_addressbase script to uk_geo_utils app
